requirements.txt
flasgger==0.9.7.1
quart==0.18.4
quart-cors==0.6.0
quart-rate-limiter==0.8.0
jsonschema==4.18.4
pre-commit==3.3.3
pydantic==2.0.3
PyJWT==2.8.0
PyMySQL==1.1.0
pymysql-pool==0.4.3
PyPika==0.48.9
python-dotenv==1.0.0
requests==2.31.0
pydantic-settings==2.0.2
